New Structure
Successfully developed a new structure for creating objects and communicating with API using EndPointBuilder and sub class Builders. Also rearranged some class locations. Primary Objective of this change is to remove excess uneeded code to focus solely on the design of the code to be applied to all other API implementations.