Huge API update
Overhauled the API system framework. Transistioned to a system that uses generics and abstraction to enable me to provide Strategy pattern for both data retrieval and endpoint building. Added more DTO objects and added support for League API (ranked). However there are still issues. Need to add support for deserializing a response from API as a set and building each individual item and return to end user as a set of data and apply it in an abstract way for supporting other APIs with similar response results.